Move randomUtxo, randomToken and randomNFT to utils
